Move Github cicd impl to github/cicd
Also move merge request state method over to
github/merge_request
